% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppUUID-package.R
\docType{package}
\name{RcppUUID}
\alias{RcppUUID}
\alias{RcppUUID-package}
\title{Generating and validating of UUIDs}
\description{
Tools for generating and validating of UUIDs (Universally Unique Identifiers).
}
\author{
\strong{Maintainer}: Artem Klevtsov \email{a.a.klevtsov@gmail.com} (\href{https://orcid.org/0000-0003-0492-6647}{ORCID})

}
