% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/uuid.R
\name{uuid_generate_random}
\alias{uuid_generate_random}
\alias{uuid_generate_nil}
\alias{uuid_generate_name}
\alias{uuid_generate}
\title{Generate UUID}
\usage{
uuid_generate_random(n = 1L)

uuid_generate_nil(n = 1L)

uuid_generate_name(x)
}
\arguments{
\item{n}{Number of generated UUIDs.}

\item{x}{Character vector.}
}
\value{
Character vector with UUIDs.
}
\description{
Function generates a set of Universally Unique Identifiers.
}
\examples{
# generate random UUIDs
uuid_generate_random(2)
# generate nil UUIDs
uuid_generate_nil(2)
# generate name UUIDs
uuid_generate_name(c("one", "two"))

}
