% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uuid_generate_name}
\alias{uuid_generate_name}
\title{Generate UUIDs Version 5}
\usage{
uuid_generate_name(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector with UUIDs.
}
\description{
Function generates uuids using operating system provided entropy.
}
\note{
This function generates valid uuids for the \code{NA}.
}
\examples{
# generate name UUIDs
uuid_generate_name(c("one", "two"))

}
\references{
\url{https://www.boost.org/doc/libs/1_72_0/libs/uuid/doc/uuid.html#Name\%20Generator}
}
