% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uuid_generate_nil}
\alias{uuid_generate_nil}
\title{Generates Nil UUIDs}
\usage{
uuid_generate_nil(n = 1L)
}
\arguments{
\item{n}{Number of generated UUIDs.}
}
\value{
Character vector with UUIDs.
}
\description{
Function generates nil uuids.
}
\examples{
# generate nil UUIDs
uuid_generate_nil(2)

}
\references{
\url{https://www.boost.org/doc/libs/1_72_0/libs/uuid/doc/uuid.html#Nil\%20Generator}
}
