% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uuid_generate_random}
\alias{uuid_generate_random}
\title{Generate UUIDs Version 4}
\usage{
uuid_generate_random(n = 1L)
}
\arguments{
\item{n}{Number of generated UUIDs.}
}
\value{
Character vector with UUIDs.
}
\description{
Function generates uuids using operating system provided entropy.
}
\examples{
# generate random UUIDs
uuid_generate_random(2)

}
\references{
\url{https://www.boost.org/doc/libs/1_72_0/libs/uuid/doc/uuid.html#Random\%20Generator}
}
