% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uuid_generate_time}
\alias{uuid_generate_time}
\title{Generate UUIDs Version 7}
\usage{
uuid_generate_time(n = 1L)
}
\arguments{
\item{n}{Number of generated UUIDs.}
}
\value{
Character vector with UUIDs generated under specification v7.
}
\description{
This function generates time-based uuids using operating system provided entropy. Note
that this requires Boost 1.86 or later.
}
\examples{
# generate random UUIDs
uuid_generate_time(2)

}
\references{
\url{https://www.boost.org/doc/libs/release/libs/uuid/}
}
