% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uuid_validate}
\alias{uuid_validate}
\title{Validate UUIDs}
\usage{
uuid_validate(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Logical vector.
}
\description{
Function validates uuids.
}
\examples{
# validate UUIDs
uuid_validate(uuid_generate_random(2))
uuid_validate(uuid_generate_nil(2))
uuid_validate(uuid_generate_name(c("one", "two")))
uuid_validate(c("a", ""))

}
