% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppUUID-package.R
\docType{package}
\name{RcppUUID}
\alias{RcppUUID}
\alias{RcppUUID-package}
\title{Generating Universally Unique Identificators}
\description{
Provides functions to generating a vector of Universally Unique Identifiers (UUID).
Used implementation from the Boost C++ library. Supported random (version 4) and name (version 5)
UUIDs. UUIDs generation are parallelized by OpenMP.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://artemklevtsov.gitlab.io/rcppuuid}
  \item \url{https://gitlab.com/artemklevtsov/rcppuuid}
  \item Report bugs at \url{https://gitlab.com/artemklevtsov/rcppuuid/issues}
}

}
\author{
\strong{Maintainer}: Artem Klevtsov \email{a.a.klevtsov@gmail.com} (\href{https://orcid.org/0000-0003-0492-6647}{ORCID})

}
