% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppUUID-package.R
\docType{package}
\name{RcppUUID}
\alias{RcppUUID}
\alias{RcppUUID-package}
\title{Generating Universally Unique Identificators}
\description{
Provides funstion to generating a vector of Universally Unique Identifiers
(UUID) version 4. Based on the sole C++ library <https://github.com/r-lyeh-archived/sole>.
Used the OpenMP library to improve performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://artemklevtsov.gitlab.io/rcppuuid}
  \item \url{https://gitlab.com/artemklevtsov/rcppuuid}
  \item Report bugs at \url{https://gitlab.com/artemklevtsov/rcppuuid/issues}
}

}
\author{
\strong{Maintainer}: Artem Klevtsov \email{a.a.klevtsov@gmail.com} (\href{https://orcid.org/0000-0003-0492-6647}{ORCID})

}
