% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uuid_generate}
\alias{uuid_generate}
\title{Generate a version 4 UUID}
\usage{
uuid_generate(n = 1L)
}
\arguments{
\item{n}{Number of generated UUIDs.}
}
\value{
Character vector with UUIDs.
}
\description{
Function generates a new Universally Unique Identifier.
}
\examples{
uuid_generate(5)

}
