% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uuid_validate}
\alias{uuid_validate}
\title{Validate UUID}
\usage{
uuid_validate(x)
}
\arguments{
\item{x}{Character vector with UUIDs.}
}
\value{
Loical vector.
}
\description{
Function validate a Universally Unique Identifiers.
}
\examples{
uuid_validate(uuid_generate(5))

}
